+++ src/shell/zxdg.rs
@@ -1,28 +1,21 @@
 use wayland_client::protocol::{wl_output, wl_seat, wl_surface};
-use wayland_client::Proxy;
 
 use wayland_protocols::unstable::xdg_shell::v6::client::{
     zxdg_shell_v6, zxdg_surface_v6, zxdg_toplevel_v6,
 };
 use wayland_protocols::xdg_shell::client::xdg_toplevel;
 
-use self::zxdg_shell_v6_proto::zxdg_shell_v6::RequestsTrait as ShellRequests;
-use self::zxdg_shell_v6_proto::zxdg_surface_v6::RequestsTrait as SurfaceRequests;
-use self::zxdg_shell_v6_proto::zxdg_toplevel_v6::RequestsTrait as ToplevelRequests;
-use wayland_client::protocol::wl_surface::RequestsTrait as WlSurfaceRequests;
-use wayland_protocols::unstable::xdg_shell::v6::client as zxdg_shell_v6_proto;
-
 use super::{Event, ShellSurface};
 
 pub(crate) struct Zxdg {
-    surface: Proxy<zxdg_surface_v6::ZxdgSurfaceV6>,
-    toplevel: Proxy<zxdg_toplevel_v6::ZxdgToplevelV6>,
+    surface: zxdg_surface_v6::ZxdgSurfaceV6,
+    toplevel: zxdg_toplevel_v6::ZxdgToplevelV6,
 }
 
 impl Zxdg {
     pub(crate) fn create<Impl>(
-        surface: &Proxy<wl_surface::WlSurface>,
-        shell: &Proxy<zxdg_shell_v6::ZxdgShellV6>,
+        surface: &wl_surface::WlSurface,
+        shell: &zxdg_shell_v6::ZxdgShellV6,
         mut implementation: Impl,
     ) -> Zxdg
     where
@@ -30,11 +23,12 @@ impl Zxdg {
     {
         let xdgs = shell
             .get_xdg_surface(surface, |xdgs| {
-                xdgs.implement(
-                    |evt, xdgs: Proxy<_>| match evt {
+                xdgs.implement_closure(
+                    |evt, xdgs| match evt {
                         zxdg_surface_v6::Event::Configure { serial } => {
                             xdgs.ack_configure(serial);
                         }
+                        _ => unreachable!(),
                     },
                     (),
                 )
@@ -42,7 +36,7 @@ impl Zxdg {
             .unwrap();
         let toplevel = xdgs
             .get_toplevel(|toplevel| {
-                toplevel.implement(
+                toplevel.implement_closure(
                     move |evt, _| {
                         match evt {
                             zxdg_toplevel_v6::Event::Close => implementation(Event::Close),
@@ -72,6 +66,7 @@ impl Zxdg {
                                     .collect::<Vec<_>>();
                                 implementation(Event::Configure { new_size, states });
                             }
+                            _ => unreachable!(),
                         }
                     },
                     (),
@@ -87,11 +82,11 @@ impl Zxdg {
 }
 
 impl ShellSurface for Zxdg {
-    fn resize(&self, seat: &Proxy<wl_seat::WlSeat>, serial: u32, edges: xdg_toplevel::ResizeEdge) {
+    fn resize(&self, seat: &wl_seat::WlSeat, serial: u32, edges: xdg_toplevel::ResizeEdge) {
         self.toplevel.resize(seat, serial, edges as u32);
     }
 
-    fn move_(&self, seat: &Proxy<wl_seat::WlSeat>, serial: u32) {
+    fn move_(&self, seat: &wl_seat::WlSeat, serial: u32) {
         self.toplevel._move(seat, serial);
     }
 
@@ -103,7 +98,7 @@ impl ShellSurface for Zxdg {
         self.toplevel.set_app_id(app_id);
     }
 
-    fn set_fullscreen(&self, output: Option<&Proxy<wl_output::WlOutput>>) {
+    fn set_fullscreen(&self, output: Option<&wl_output::WlOutput>) {
         self.toplevel.set_fullscreen(output)
     }
 
@@ -143,7 +138,7 @@ impl ShellSurface for Zxdg {
         }
     }
 
-    fn get_xdg(&self) -> Option<&Proxy<xdg_toplevel::XdgToplevel>> {
+    fn get_xdg(&self) -> Option<&xdg_toplevel::XdgToplevel> {
         None
     }
 }