+++ src/surface.rs
@@ -2,21 +2,19 @@
 use env::Environment;
 use output::OutputMgr;
 use std::sync::Mutex;
-use wayland_client::protocol::wl_compositor::RequestsTrait as CompositorRequest;
 use wayland_client::protocol::{wl_output, wl_surface};
-use wayland_client::Proxy;
 
 pub(crate) struct SurfaceUserData {
     dpi_factor: i32,
-    outputs: Vec<Proxy<wl_output::WlOutput>>,
+    outputs: Vec<wl_output::WlOutput>,
     output_manager: OutputMgr,
-    dpi_change_cb: Box<FnMut(i32, Proxy<wl_surface::WlSurface>) + Send + 'static>,
+    dpi_change_cb: Box<FnMut(i32, wl_surface::WlSurface) + Send + 'static>,
 }
 
 impl SurfaceUserData {
     fn new(
         output_manager: OutputMgr,
-        dpi_change_cb: Box<FnMut(i32, Proxy<wl_surface::WlSurface>) + Send + 'static>,
+        dpi_change_cb: Box<FnMut(i32, wl_surface::WlSurface) + Send + 'static>,
     ) -> Self {
         SurfaceUserData {
             dpi_factor: 1,
@@ -26,25 +24,18 @@ impl SurfaceUserData {
         }
     }
 
-    pub(crate) fn enter(
-        &mut self,
-        output: Proxy<wl_output::WlOutput>,
-        surface: Proxy<wl_surface::WlSurface>,
-    ) {
+    pub(crate) fn enter(&mut self, output: wl_output::WlOutput, surface: wl_surface::WlSurface) {
         self.outputs.push(output);
         self.compute_dpi_factor(surface);
     }
 
-    pub(crate) fn leave(
-        &mut self,
-        output: &Proxy<wl_output::WlOutput>,
-        surface: Proxy<wl_surface::WlSurface>,
-    ) {
-        self.outputs.retain(|output2| !output.equals(output2));
+    pub(crate) fn leave(&mut self, output: &wl_output::WlOutput, surface: wl_surface::WlSurface) {
+        self.outputs
+            .retain(|output2| !output.as_ref().equals(output2.as_ref()));
         self.compute_dpi_factor(surface);
     }
 
-    fn compute_dpi_factor(&mut self, surface: Proxy<wl_surface::WlSurface>) {
+    fn compute_dpi_factor(&mut self, surface: wl_surface::WlSurface) {
         let mut scale_factor = 1;
         for output in &self.outputs {
             if let Some(scale_factor2) = self
@@ -64,22 +55,22 @@ impl SurfaceUserData {
 /// Creates a WlSurface from an Environment.
 ///
 /// Computes the dpi factor by taking the maximum dpi value of all the outputs
-/// the surface is displayed on. The dpi factor is stored in the Proxy's user
-/// data. When the dpi value is updated the caller is notified through the
-/// dpi_change closure.
+/// the surface is displayed on. When the dpi value is updated the caller is
+/// notified through the dpi_change closure.
 pub(crate) fn create_surface<F>(
     environment: &Environment,
     dpi_change: Box<F>,
-) -> Proxy<wl_surface::WlSurface>
+) -> wl_surface::WlSurface
 where
-    F: FnMut(i32, Proxy<wl_surface::WlSurface>) + Send + 'static,
+    F: FnMut(i32, wl_surface::WlSurface) + Send + 'static,
 {
     environment
         .compositor
         .create_surface(move |surface| {
-            surface.implement(
+            surface.implement_closure(
                 move |event, surface| {
                     let mut user_data = surface
+                        .as_ref()
                         .user_data::<Mutex<SurfaceUserData>>()
                         .unwrap()
                         .lock()
@@ -91,6 +82,7 @@ where
                         wl_surface::Event::Leave { output } => {
                             user_data.leave(&output, surface.clone());
                         }
+                        _ => unreachable!(),
                     };
                 },
                 Mutex::new(SurfaceUserData::new(
@@ -103,8 +95,9 @@ where
 }
 
 /// Returns the current dpi factor of a surface.
-pub fn get_dpi_factor(surface: &Proxy<wl_surface::WlSurface>) -> i32 {
+pub fn get_dpi_factor(surface: &wl_surface::WlSurface) -> i32 {
     surface
+        .as_ref()
         .user_data::<Mutex<SurfaceUserData>>()
         .expect("Surface was not created with create_surface.")
         .lock()
@@ -113,8 +106,9 @@ pub fn get_dpi_factor(surface: &Proxy<wl_surface::WlSurface>) -> i32 {
 }
 
 /// Returns a list of outputs the surface is displayed on.
-pub fn get_outputs(surface: &Proxy<wl_surface::WlSurface>) -> Vec<Proxy<wl_output::WlOutput>> {
+pub fn get_outputs(surface: &wl_surface::WlSurface) -> Vec<wl_output::WlOutput> {
     surface
+        .as_ref()
         .user_data::<Mutex<SurfaceUserData>>()
         .expect("Surface was not created with create_surface.")
         .lock()