+++ CHANGELOG.md
@@ -2,9 +2,15 @@
 
 ## Unreleased
 
+#### Breaking changes
+
+- Update the crate to `wayland-client` version 0.22
+- Window: `set_title()` now requires a manual `refresh()` for the change to take effect
+
+#### Bugfixes
+
 - Keyboard: fix system repeat rate as repeats per second rather then millisecond delay between repeats
 - Surface: fix panic in `compute_dpi_factor()` by only computing the dpi factor on surfaces known to the OutputMgr
-- Window: `set_title()` now requires a manual `refresh()` for the change to take effect
 
 ## 0.4.4 -- 2018-12-27